Refactor: Organize codebase into multiple header and source files
I've separated the monolithic UNIFIGHT.C into a more modular structure:

- `structs.h`: Contains all structure and enum definitions.
- `defines.h`: Contains all global #define constants.
- `personagem.h`/`.c`: Manages character-related logic, including movement, animation, and combat.
- `ui.h`/`.c`: Handles UI drawing functions for menus, HUD, and selection screens.
- `efeitos.h`/`.c`: Manages particle effects.
- `config.h`/`.c`: Handles game configuration, settings, and input mapping.
- `unifight.c`: Now serves as the main program entry point, containing the game loop and including the new modules.

This refactoring should improve code organization, readability, and maintainability for you.
