+++ UNIFLIGHT-main/UNIFLIGHT-main/unifight.c
@@ -3,1017 +3,15 @@
 #include <string.h>
 #include <math.h>
 #include <locale.h>
-#include "structs.c"
-
-#define LARGURA_TELA 1920
-#define ALTURA_TELA 1080
-#define MAX_PERSONAGENS 6
-#define OPCOES_MENU 3
-#define OPCOES_CONFIG 8
-#define MAX_MAPAS 6
-#define VIDA_MAXIMA 250
-#define PODER_MAXIMO 100
-
-#define TEMPO_ROUND 60.0f
-#define MAX_ROUNDS 3
-
-// ===== FUNÇÕES DE MOVIMENTO E SPRITES =====
-
-void InicializarPosicaoPersonagem(Personagem *p, float x, float y, bool viradoDireita)
-{
-    p->posicao.x = x;
-    p->posicao.y = y;
-    p->velocidade.x = 0;
-    p->velocidade.y = 0;
-    p->velocidadeMaxima = 200.0f;
-    p->viradoParaDireita = viradoDireita;
-    p->podeAtacar = true;
-    p->podeSeMovimentar = true;
-
-    // Configurar hitbox
-    p->hitbox.width = 80;
-    p->hitbox.height = 220; // Aumentado para cobrir mais do personagem
-    p->hitbox.x = p->posicao.x - p->hitbox.width / 2;
-    p->hitbox.y = p->posicao.y - p->hitbox.height;
-
-    // Configurar alcances de ataque
-    p->alcanceSoco = 90.0f;
-    p->alcanceChute = 110.0f;
-    p->alcancePoder = 150.0f;
-
-    // Sistema de sprites melhorado
-    p->frameAtual = 0;
-    p->totalFrames = 4;
-    p->timerFrame = 0.0f;
-    p->duracaoFrame = 0.12f;                    // Animação mais rápida
-    p->frameSource = (Rectangle){0, 0, 64, 64}; // Tamanho padrão do frame
-}
-
-void AtualizarHitbox(Personagem *p)
-{
-    p->hitbox.x = p->posicao.x - p->hitbox.width / 2;
-    p->hitbox.y = p->posicao.y - p->hitbox.height;
-}
-
-void AtualizarAlcanceAtaque(Personagem *p)
-{
-    float alcance = 0;
-    switch (p->poseAtual)
-    {
-    case POSE_SOCO:
-        alcance = p->alcanceSoco;
-        break;
-    case POSE_CHUTE:
-        alcance = p->alcanceChute;
-        break;
-    case POSE_PODER:
-        alcance = p->alcancePoder;
-        break;
-    default:
-        alcance = 0;
-        break;
-    }
-
-    if (p->viradoParaDireita)
-    {
-        p->alcanceAtaque.x = p->posicao.x;
-        p->alcanceAtaque.y = p->posicao.y - 60;
-        p->alcanceAtaque.width = alcance;
-        p->alcanceAtaque.height = 80;
-    }
-    else
-    {
-        p->alcanceAtaque.x = p->posicao.x - alcance;
-        p->alcanceAtaque.y = p->posicao.y - 60;
-        p->alcanceAtaque.width = alcance;
-        p->alcanceAtaque.height = 80;
-    }
-}
-
-bool VerificarProximidade(Personagem *atacante, Personagem *alvo)
-{
-    AtualizarAlcanceAtaque(atacante);
-    return CheckCollisionRecs(atacante->alcanceAtaque, alvo->hitbox);
-}
-
-void MoverPersonagem(Personagem *p, float deltaTime, float limiteEsquerdo, float limiteDireito)
-{
-    if (!p->podeSeMovimentar || p->animando)
-        return;
-
-    // Aplicar movimento
-    p->posicao.x += p->velocidade.x * deltaTime;
-
-    // Verificar limites da arena
-    if (p->posicao.x < limiteEsquerdo + p->hitbox.width / 2)
-    {
-        p->posicao.x = limiteEsquerdo + p->hitbox.width / 2;
-        p->velocidade.x = 0;
-    }
-    if (p->posicao.x > limiteDireito - p->hitbox.width / 2)
-    {
-        p->posicao.x = limiteDireito - p->hitbox.width / 2;
-        p->velocidade.x = 0;
-    }
-
-    // Atualizar hitbox
-    AtualizarHitbox(p);
-
-    // Determinar direção que está virado
-    if (p->velocidade.x > 0)
-        p->viradoParaDireita = true;
-    else if (p->velocidade.x < 0)
-        p->viradoParaDireita = false;
-}
-
-void ProcessarMovimentoPlayer1(Personagem *p, ConfiguracaoJogo *config)
-{
-    if (!p->podeSeMovimentar || p->animando)
-        return;
-
-    p->velocidade.x = 0;
-    bool movendo = false;
-
-    // Movimento mais responsivo estilo Street Fighter
-    if (IsKeyDown(config->teclaEsquerdaP1))
-    {
-        p->velocidade.x = -p->velocidadeMaxima;
-        movendo = true;
-    }
-    else if (IsKeyDown(config->teclaDireitaP1))
-    {
-        p->velocidade.x = p->velocidadeMaxima;
-        movendo = true;
-    }
-
-    // Atualizar pose baseada no movimento
-    if (movendo && p->poseAtual == POSE_IDLE)
-    {
-        p->poseAtual = POSE_WALK;
-    }
-    else if (!movendo && p->poseAtual == POSE_WALK && !p->animando)
-    {
-        p->poseAtual = POSE_IDLE;
-    }
-}
-
-void ProcessarMovimentoPlayer2(Personagem *p, ConfiguracaoJogo *config)
-{
-    if (!p->podeSeMovimentar || p->animando)
-        return;
-
-    p->velocidade.x = 0;
-    bool movendo = false;
-
-    // Movimento mais responsivo estilo Street Fighter
-    if (IsKeyDown(config->teclaEsquerdaP2))
-    {
-        p->velocidade.x = -p->velocidadeMaxima;
-        movendo = true;
-    }
-    else if (IsKeyDown(config->teclaDireitaP2))
-    {
-        p->velocidade.x = p->velocidadeMaxima;
-        movendo = true;
-    }
-
-    // Atualizar pose baseada no movimento
-    if (movendo && p->poseAtual == POSE_IDLE)
-    {
-        p->poseAtual = POSE_WALK;
-    }
-    else if (!movendo && p->poseAtual == POSE_WALK && !p->animando)
-    {
-        p->poseAtual = POSE_IDLE;
-    }
-}
-
-void AtualizarAnimacaoSprite(Personagem *p, float deltaTime)
-{
-    // Atualizar timer do frame
-    p->timerFrame += deltaTime;
-
-    if (p->timerFrame >= p->duracaoFrame)
-    {
-        p->timerFrame = 0.0f;
-        p->frameAtual++;
-
-        // Configurar frames por pose com animações mais suaves
-        int maxFrames = 1;
-        switch (p->poseAtual)
-        {
-        case POSE_IDLE:
-            maxFrames = 4; // Respiração suave
-            p->duracaoFrame = 0.25f;
-            break;
-        case POSE_WALK:
-            maxFrames = 6; // Caminhada fluida
-            p->duracaoFrame = 0.15f;
-            break;
-        case POSE_SOCO:
-            maxFrames = 4; // Soco rápido
-            p->duracaoFrame = 0.08f;
-            break;
-        case POSE_CHUTE:
-            maxFrames = 5; // Chute com preparação
-            p->duracaoFrame = 0.1f;
-            break;
-        case POSE_PODER:
-            maxFrames = 8; // Poder especial elaborado
-            p->duracaoFrame = 0.12f;
-            break;
-        case POSE_DANO:
-            maxFrames = 3; // Reação ao dano
-            p->duracaoFrame = 0.15f;
-            break;
-        case POSE_DEFESA:
-            maxFrames = 2; // Posição defensiva
-            p->duracaoFrame = 0.2f;
-            break;
-        }
-
-        if (p->frameAtual >= maxFrames)
-        {
-            p->frameAtual = 0;
-
-            // Se é uma animação de ataque, voltar ao idle
-            if (p->animando && (p->poseAtual == POSE_SOCO || p->poseAtual == POSE_CHUTE ||
-                                p->poseAtual == POSE_PODER || p->poseAtual == POSE_DANO))
-            {
-                // A animação será finalizada pela função AtualizarAnimacao
-            }
-        }
-    }
-
-    // Configurar source rectangle para sprite
-    int frameWidth = 64;
-    int frameHeight = 64;
-
-    p->frameSource.x = p->frameAtual * frameWidth;
-    p->frameSource.y = (int)p->poseAtual * frameHeight;
-    p->frameSource.width = frameWidth;
-    p->frameSource.height = frameHeight;
-}
-
-Texture2D ObterTexturaAtual(Personagem *p)
-{
-    switch (p->poseAtual)
-    {
-    case POSE_WALK:
-        return p->texturaWalk.id > 0 ? p->texturaWalk : p->texturaLuta;
-    case POSE_SOCO:
-        return p->texturaSoco.id > 0 ? p->texturaSoco : p->texturaLuta;
-    case POSE_CHUTE:
-        return p->texturaChute.id > 0 ? p->texturaChute : p->texturaLuta;
-    case POSE_PODER:
-        return p->texturaPoder.id > 0 ? p->texturaPoder : p->texturaLuta;
-    case POSE_DEFESA:
-        return p->texturaDefesa.id > 0 ? p->texturaDefesa : p->texturaLuta;
-    case POSE_DANO:
-        return p->texturaDano.id > 0 ? p->texturaDano : p->texturaLuta;
-    case POSE_IDLE:
-    default:
-        return p->texturaLuta;
-    }
-}
-
-void IniciarAnimacao(Personagem *p, TipoPose pose, float duracao)
-{
-    p->poseAtual = pose;
-    p->timerAnimacao = 0.0f;
-    p->duracaoAnimacao = duracao;
-    p->animando = true;
-    p->frameAtual = 0; // Resetar frame para início da animação
-    p->timerFrame = 0.0f;
-
-    // Bloquear movimento durante ataques
-    if (pose == POSE_SOCO || pose == POSE_CHUTE || pose == POSE_PODER)
-    {
-        p->podeSeMovimentar = false;
-        p->podeAtacar = false;
-    }
-}
-
-void AtualizarAnimacao(Personagem *p, float deltaTime)
-{
-    if (p->animando)
-    {
-        p->timerAnimacao += deltaTime;
-        if (p->timerAnimacao >= p->duracaoAnimacao)
-        {
-            p->poseAtual = POSE_IDLE;
-            p->animando = false;
-            p->timerAnimacao = 0.0f;
-            p->frameAtual = 0;
-            p->timerFrame = 0.0f;
-
-            // Restaurar capacidades
-            p->podeSeMovimentar = true;
-            p->podeAtacar = true;
-        }
-    }
-}
-
-Rectangle ObterSourceRectSprite(Personagem *p)
-{
-    return p->frameSource;
-}
-
-void VirarPersonagemParaOponente(Personagem *p1, Personagem *p2)
-{
-    // Só virar se não estiver atacando
-    if (!p1->animando || (p1->poseAtual != POSE_SOCO && p1->poseAtual != POSE_CHUTE && p1->poseAtual != POSE_PODER))
-    {
-        if (p2->posicao.x > p1->posicao.x)
-            p1->viradoParaDireita = true;
-        else
-            p1->viradoParaDireita = false;
-    }
-
-    if (!p2->animando || (p2->poseAtual != POSE_SOCO && p2->poseAtual != POSE_CHUTE && p2->poseAtual != POSE_PODER))
-    {
-        if (p1->posicao.x > p2->posicao.x)
-            p2->viradoParaDireita = true;
-        else
-            p2->viradoParaDireita = false;
-    }
-}
-
-bool ExecutarAtaque(Personagem *atacante, Personagem *alvo, TipoPose tipoAtaque, int dano)
-{
-    if (!atacante->podeAtacar || atacante->animando)
-        return false;
-
-    // Verificar se está próximo o suficiente
-    if (!VerificarProximidade(atacante, alvo))
-        return false;
-
-    // Executar ataque
-    alvo->vidaAtual -= dano;
-    if (alvo->vidaAtual < 0)
-        alvo->vidaAtual = 0;
-
-    // Iniciar animações
-    float duracaoAtaque = 0.4f;
-    if (tipoAtaque == POSE_SOCO)
-        duracaoAtaque = 0.3f;
-    else if (tipoAtaque == POSE_CHUTE)
-        duracaoAtaque = 0.5f;
-    else if (tipoAtaque == POSE_PODER)
-        duracaoAtaque = 0.8f;
-
-    IniciarAnimacao(atacante, tipoAtaque, duracaoAtaque);
-    IniciarAnimacao(alvo, POSE_DANO, 0.3f);
-
-    // Empurrar o alvo para trás
-    float forcaEmpurrao = 30.0f;
-    if (tipoAtaque == POSE_CHUTE)
-        forcaEmpurrao = 45.0f;
-    else if (tipoAtaque == POSE_PODER)
-        forcaEmpurrao = 60.0f;
-
-    if (atacante->viradoParaDireita)
-    {
-        alvo->posicao.x += forcaEmpurrao;
-    }
-    else
-    {
-        alvo->posicao.x -= forcaEmpurrao;
-    }
-
-    return true;
-}
-
-void DesenharSelecaoPersonagem(Personagem personagens[], int selecionado, int inicioX, int inicioY, int larguraArea, int alturaArea, Color corDestaque)
-{
-    int espacamento = 35;
-    int larguraPers = 150;
-    int alturaPers = 150;
-
-    int colunas = 3;
-    int linhas = 2;
-
-    int larguraTotal = colunas * larguraPers + (colunas - 1) * espacamento;
-    int alturaTotal = linhas * alturaPers + (linhas - 1) * espacamento;
-
-    int posXInicial = inicioX + (larguraArea - larguraTotal) / 2;
-    int posYInicial = inicioY + (alturaArea - alturaTotal) / 2;
-
-    for (int i = 0; i < MAX_PERSONAGENS; i++)
-    {
-        int linha = i / colunas;
-        int coluna = i % colunas;
-
-        int x = posXInicial + coluna * (larguraPers + espacamento);
-        int y = posYInicial + linha * (alturaPers + espacamento);
-
-        Rectangle destino = {(float)x, (float)y, (float)larguraPers, (float)alturaPers};
-        Color cor = WHITE;
-
-        if (i == selecionado)
-        {
-            destino.width *= 1.2f;
-            destino.height *= 1.2f;
-            DrawRectangleLines(x - 5, y - 5, (int)(larguraPers * 1.2f) + 10, (int)(alturaPers * 1.2f) + 10, corDestaque);
-        }
-
-        DrawTexturePro(personagens[i].textura,
-                       (Rectangle){0, 0, (float)personagens[i].textura.width, (float)personagens[i].textura.height},
-                       destino,
-                       (Vector2){0, 0},
-                       0.0f,
-                       cor);
-    }
-}
-
-void DesenharInfoPersonagem(Personagem personagem, int x, int y, int largura, int altura, Font fonte)
-{
-    // Painel de fundo com gradiente
-    DrawRectangleRounded((Rectangle){x, y, largura, altura}, 0.1f, 10, Fade(BLACK, 0.85f));
-    DrawRectangleRoundedLines((Rectangle){x, y, largura, altura}, 0.1f, 10, personagem.corHabilidade);
-
-    int posY = y + 25;
-    int espacamento = 45;
-
-    // Nome do personagem com fonte ainda maior
-    DrawTextEx(fonte, personagem.nome, (Vector2){x + 25, posY}, 42, 2, WHITE);
-    posY += espacamento + 20;
-
-    // Habilidade com fonte maior e melhor destaque
-    const char *habilidadeTexto = TextFormat("Habilidade: %s", personagem.habilidade);
-    DrawTextEx(fonte, habilidadeTexto, (Vector2){x + 25, posY}, 32, 2, personagem.corHabilidade);
-    posY += espacamento;
-
-    // História com fonte maior
-    const char *historiaTexto = "História:";
-    DrawTextEx(fonte, historiaTexto, (Vector2){x + 25, posY}, 28, 2, LIGHTGRAY);
-    posY += 35;
-
-    // Texto da história com fonte maior e melhor formatação
-    int larguraTexto = largura - 50;
-    const char *historia = personagem.historia;
-    int tamanhoFonte = 24; // Aumentado de 20 para 24
-
-    char linha[250]; // Aumentado o buffer
-    int posicao = 0;
-    int caracteresPorLinha = (larguraTexto / (tamanhoFonte * 0.55f)); // Ajustado para fonte maior
-
-    if (caracteresPorLinha <= 0)
-        caracteresPorLinha = 25; // Aumentado
-
-    while (historia[posicao] != '\0' && posY < y + altura - 35)
-    {
-        int fim = posicao + caracteresPorLinha;
-        int tamanhoHistoria = strlen(historia);
-
-        if (fim >= tamanhoHistoria)
-        {
-            fim = tamanhoHistoria;
-        }
-        else
-        {
-            // Procurar quebra de palavra mais inteligente
-            while (fim > posicao && historia[fim] != ' ' && historia[fim] != '.' && historia[fim] != ',')
-            {
-                fim--;
-            }
-            // Se não encontrou espaço, quebra na palavra mesmo
-            if (fim == posicao)
-            {
-                fim = posicao + caracteresPorLinha;
-            }
-        }
-
-        int tamanho = fim - posicao;
-        if (tamanho > 0 && tamanho < 249)
-        {
-            strncpy(linha, &historia[posicao], tamanho);
-            linha[tamanho] = '\0';
-
-            // Remover espaços no início da linha
-            char *linhaLimpa = linha;
-            while (*linhaLimpa == ' ')
-                linhaLimpa++;
-
-            DrawTextEx(fonte, linhaLimpa, (Vector2){x + 25, posY}, tamanhoFonte, 2, WHITE);
-            posY += 30; // Aumentado o espaçamento entre linhas
-            posicao = (fim < tamanhoHistoria && (historia[fim] == ' ' || historia[fim] == '.' || historia[fim] == ',')) ? fim + 1 : fim;
-        }
-        else
-        {
-            break;
-        }
-    }
-}
-
-void DesenharBarraVida(int x, int y, int largura, int altura, int vidaAtual, int vidaMaxima, Color cor, const char *nomeJogador)
-{
-    // Fundo da barra com gradiente
-    DrawRectangleGradientV(x, y, largura, altura, Fade(DARKGRAY, 0.8f), DARKGRAY);
-
-    // Barra de vida com gradiente baseado na porcentagem
-    float porcentagem = (float)vidaAtual / (float)vidaMaxima;
-    int larguraVida = (int)(largura * porcentagem);
-
-    // Cor da vida baseada na porcentagem com gradiente
-    Color corVida;
-    if (porcentagem > 0.6f)
-        corVida = ColorLerp(YELLOW, GREEN, (porcentagem - 0.6f) / 0.4f);
-    else if (porcentagem > 0.3f)
-        corVida = ColorLerp(RED, ORANGE, (porcentagem - 0.3f) / 0.3f);
-    else
-        corVida = ColorLerp((Color){128, 0, 0, 255}, RED, porcentagem / 0.3f); // Substituindo DARKRED por cor RGB
-
-    // Desenhar barra com efeito de gradiente
-    DrawRectangleGradientH(x, y, larguraVida, altura, corVida, Fade(corVida, 0.7f));
-
-    // Efeito de brilho na barra
-    if (porcentagem > 0.8f)
-    {
-        float brilho = 0.3f + 0.2f * sin(GetTime() * 6);
-        DrawRectangleGradientH(x, y, larguraVida, altura / 3, Fade(WHITE, brilho), BLANK);
-    }
-
-    // Borda estilizada
-    DrawRectangleLines(x - 1, y - 1, largura + 2, altura + 2, BLACK);
-    DrawRectangleLines(x, y, largura, altura, WHITE);
-
-    // Texto da vida centralizado e estilizado
-    const char *textoVida = TextFormat("%d/%d", vidaAtual, vidaMaxima);
-    int larguraTexto = MeasureText(textoVida, 16);
-
-    // Sombra do texto
-    DrawText(textoVida, x + largura / 2 - larguraTexto / 2 + 1, y + altura / 2 - 8 + 1, 16, BLACK);
-    // Texto principal
-    DrawText(textoVida, x + largura / 2 - larguraTexto / 2, y + altura / 2 - 8, 16, WHITE);
-
-    // Nome do jogador acima da barra
-    int larguraNome = MeasureText(nomeJogador, 20);
-    DrawText(nomeJogador, x + largura / 2 - larguraNome / 2 + 1, y - 26, 20, BLACK); // Sombra
-    DrawText(nomeJogador, x + largura / 2 - larguraNome / 2, y - 25, 20, cor);
-}
-
-void DesenharBarraPoder(int x, int y, int largura, int altura, int poderAtual, int poderMaximo, Color corPersonagem)
-{
-    // Fundo da barra
-    DrawRectangleGradientV(x, y, largura, altura, Fade(DARKGRAY, 0.8f), DARKGRAY);
-
-    // Barra de poder com efeito pulsante quando disponível
-    float porcentagem = (float)poderAtual / (float)poderMaximo;
-    int larguraPoder = (int)(largura * porcentagem);
-
-    // Cor do poder com efeito especial
-    Color corPoder;
-    if (poderAtual >= 50)
-    {
-        // Efeito pulsante quando especial está disponível
-        float pulso = 0.6f + 0.4f * sin(GetTime() * 10);
-        corPoder = ColorLerp(BLUE, YELLOW, pulso);
-
-        // Efeito de partículas visuais
-        for (int i = 0; i < 3; i++)
-        {
-            float offset = i * 2.0f;
-            float alpha = 0.3f * sin(GetTime() * 8 + offset);
-            DrawRectangle(x + larguraPoder - 5, y - 2, 10, altura + 4, Fade(YELLOW, alpha));
-        }
-    }
-    else
-    {
-        corPoder = ColorLerp(DARKBLUE, BLUE, porcentagem);
-    }
-
-    DrawRectangleGradientH(x, y, larguraPoder, altura, corPoder, Fade(corPoder, 0.6f));
-
-    // Linha indicadora dos 50 pontos (metade da barra)
-    int linhaMeio = x + largura / 2;
-    DrawLine(linhaMeio, y, linhaMeio, y + altura, Fade(WHITE, 0.8f));
-    DrawLine(linhaMeio - 1, y, linhaMeio - 1, y + altura, Fade(BLACK, 0.5f));
-
-    // Borda estilizada
-    DrawRectangleLines(x - 1, y - 1, largura + 2, altura + 2, BLACK);
-    DrawRectangleLines(x, y, largura, altura, WHITE);
-
-    // Texto do poder
-    const char *textoPoder = TextFormat("%d/%d", poderAtual, poderMaximo);
-    int larguraTexto = MeasureText(textoPoder, 14);
-
-    // Sombra do texto
-    DrawText(textoPoder, x + largura / 2 - larguraTexto / 2 + 1, y + altura / 2 - 7 + 1, 14, BLACK);
-    DrawText(textoPoder, x + largura / 2 - larguraTexto / 2, y + altura / 2 - 7, 14, WHITE);
-
-    // Indicador de poder especial disponível
-    if (poderAtual >= 50)
-    {
-        const char *textoEspecial = "ESPECIAL PRONTO!";
-        float alpha = 0.7f + 0.3f * sin(GetTime() * 6);
-        Color corTexto = ColorLerp(YELLOW, WHITE, alpha);
-
-        DrawText(textoEspecial, x + largura + 15 + 1, y + altura / 2 - 7 + 1, 14, BLACK); // Sombra
-        DrawText(textoEspecial, x + largura + 15, y + altura / 2 - 7, 14, corTexto);
-    }
-}
-
-void DesenharHUD(EstadoJogo *estado, Font fonte)
-{
-    // Verificar se os jogadores existem
-    if (!estado->jogador1 || !estado->jogador2)
-        return;
-
-    // Painel superior para HUD
-    DrawRectangleGradientV(0, 0, LARGURA_TELA, 180, Fade(BLACK, 0.7f), Fade(BLACK, 0.3f));
-    DrawLine(0, 180, LARGURA_TELA, 180, Fade(WHITE, 0.5f));
-
-    // === PLAYER 1 HUD (lado esquerdo) ===
-    int margemEsquerda = 30;
-    int larguraBarra = 350;
-    int alturaBarra = 25;
-
-    DesenharBarraVida(margemEsquerda, 45, larguraBarra, alturaBarra,
-                      estado->jogador1->vidaAtual, estado->jogador1->vidaMaxima,
-                      BLUE, estado->jogador1->nome);
-
-    // Barra de poder do Player 1
-    DesenharBarraPoder(margemEsquerda, 85, larguraBarra, 20,
-                       estado->jogador1->poderAtual, estado->jogador1->poderMaximo,
-                       estado->jogador1->corHabilidade);
-
-    // Habilidade do Player 1
-    const char *habilidadeP1 = TextFormat("Habilidade: %s", estado->jogador1->habilidade);
-    DrawTextEx(fonte, habilidadeP1, (Vector2){margemEsquerda, 115}, 18, 2, estado->jogador1->corHabilidade);
-
-    // === PLAYER 2 HUD (lado direito) ===
-    int margemDireita = LARGURA_TELA - larguraBarra - 30;
-
-    DesenharBarraVida(margemDireita, 45, larguraBarra, alturaBarra,
-                      estado->jogador2->vidaAtual, estado->jogador2->vidaMaxima,
-                      RED, estado->jogador2->nome);
-
-    // Barra de poder do Player 2
-    DesenharBarraPoder(margemDireita, 85, larguraBarra, 20,
-                       estado->jogador2->poderAtual, estado->jogador2->poderMaximo,
-                       estado->jogador2->corHabilidade);
-
-    // Habilidade do Player 2
-    const char *habilidadeP2 = TextFormat("Habilidade: %s", estado->jogador2->habilidade);
-    Vector2 tamanhoHab2 = MeasureTextEx(fonte, habilidadeP2, 18, 2);
-    DrawTextEx(fonte, habilidadeP2, (Vector2){margemDireita + larguraBarra - tamanhoHab2.x, 115}, 18, 2, estado->jogador2->corHabilidade);
-
-    // === INFORMAÇÕES CENTRAIS ===
-    // Round atual
-    const char *textoRound = TextFormat("ROUND %d/%d", estado->roundAtual, MAX_ROUNDS);
-    Vector2 tamanhoRound = MeasureTextEx(fonte, textoRound, 32, 2);
-    DrawTextEx(fonte, textoRound, (Vector2){LARGURA_TELA / 2 - tamanhoRound.x / 2, 25}, 32, 2, YELLOW);
-
-    // Timer com efeito de urgência
-    Color corTempo = WHITE;
-    if (estado->tempoRound <= 10.0f)
-    {
-        float intensidade = 0.5f + 0.5f * sin(GetTime() * 8);
-        corTempo = ColorLerp(WHITE, RED, intensidade);
-    }
-
-    const char *textoTempo = TextFormat("%.0f", estado->tempoRound);
-    Vector2 tamanhoTempo = MeasureTextEx(fonte, textoTempo, 48, 2);
-    // Sombra do tempo
-    DrawTextEx(fonte, textoTempo, (Vector2){LARGURA_TELA / 2 - tamanhoTempo.x / 2 + 2, 62}, 48, 2, BLACK);
-    DrawTextEx(fonte, textoTempo, (Vector2){LARGURA_TELA / 2 - tamanhoTempo.x / 2, 60}, 48, 2, corTempo);
-
-    // Indicador VS entre os jogadores
-    const char *textoVS = "VS";
-    Vector2 tamanhoVS = MeasureTextEx(fonte, textoVS, 24, 2);
-    DrawTextEx(fonte, textoVS, (Vector2){LARGURA_TELA / 2 - tamanhoVS.x / 2, 135}, 24, 2, Fade(WHITE, 0.8f));
-}
-
-void DesenharTelaSelecaoMapa(Mapa mapas[], int mapaAtual, Font fonte)
-{
-
-    DrawRectangle(0, 0, LARGURA_TELA, ALTURA_TELA, Fade(BLACK, 0.8f));
-
-    const char *titulo = "ESCOLHA O MAPA DE LUTA";
-    Vector2 tamanhoTitulo = MeasureTextEx(fonte, titulo, 52, 2);
-    DrawTextEx(fonte, titulo, (Vector2){LARGURA_TELA / 2 - tamanhoTitulo.x / 2, 100}, 52, 2, WHITE);
-
-    int larguraMapa = 800;
-    int alturaMapa = 450;
-    int xMapa = (LARGURA_TELA - larguraMapa) / 2;
-    int yMapa = 200;
-
-    DrawTexturePro(mapas[mapaAtual].textura,
-                   (Rectangle){0, 0, mapas[mapaAtual].textura.width, mapas[mapaAtual].textura.height},
-                   (Rectangle){xMapa, yMapa, larguraMapa, alturaMapa},
-                   (Vector2){0, 0}, 0.0f, WHITE);
-
-    DrawRectangleLines(xMapa - 5, yMapa - 5, larguraMapa + 10, alturaMapa + 10, YELLOW);
-
-    const char *nomeMapa = mapas[mapaAtual].nome;
-    Vector2 tamanhoNome = MeasureTextEx(fonte, nomeMapa, 40, 2);
-    DrawTextEx(fonte, nomeMapa, (Vector2){LARGURA_TELA / 2 - tamanhoNome.x / 2, yMapa + alturaMapa + 30}, 40, 2, YELLOW);
-
-    const char *descricao = mapas[mapaAtual].descricao;
-    Vector2 tamanhoDesc = MeasureTextEx(fonte, descricao, 28, 2);
-    DrawTextEx(fonte, descricao, (Vector2){LARGURA_TELA / 2 - tamanhoDesc.x / 2, yMapa + alturaMapa + 80}, 28, 2, LIGHTGRAY);
-
-    DrawTextEx(fonte, "← → para navegar mapas", (Vector2){50, ALTURA_TELA - 100}, 28, 2, WHITE);
-    DrawTextEx(fonte, "ENTER para confirmar", (Vector2){50, ALTURA_TELA - 70}, 28, 2, GREEN);
-    DrawTextEx(fonte, "BACKSPACE para voltar", (Vector2){50, ALTURA_TELA - 40}, 28, 2, GRAY);
-}
-
-void InicializarPersonagem(Personagem *p, const char *nomePersonagem)
-{
-    p->vidaMaxima = VIDA_MAXIMA;
-    p->vidaAtual = VIDA_MAXIMA;
-    p->poderMaximo = PODER_MAXIMO;
-    p->poderAtual = PODER_MAXIMO; // Começar com poder no máximo
-    p->danoSoco = 20;
-    p->danoChute = 30;
-    p->danoPoder = 50;
-
-    p->poseAtual = POSE_IDLE;
-    p->timerAnimacao = 0.0f;
-    p->duracaoAnimacao = 0.0f;
-    p->animando = false;
-
-    // Inicializar sistema de movimento e sprites
-    p->velocidade = (Vector2){0, 0};
-    p->velocidadeMaxima = 200.0f;
-    p->viradoParaDireita = true;
-    p->podeAtacar = true;
-    p->podeSeMovimentar = true;
-
-    // Inicializar sprites
-    p->frameAtual = 0;
-    p->totalFrames = 4;
-    p->timerFrame = 0.0f;
-    p->duracaoFrame = 0.12f;
-    p->frameSource = (Rectangle){0, 0, 64, 64};
-
-    p->texturaDefesa = p->texturaLuta;
-    p->texturaDano = p->texturaLuta;
-}
-
-void CriarEfeitoImpacto(SistemaParticulas *sistema, Vector2 posicao, Color cor, int numParticulas)
-{
-    sistema->ativo = true;
-    sistema->timer = 0.0f;
-    sistema->numParticulas = numParticulas;
-
-    for (int i = 0; i < numParticulas && i < 50; i++)
-    {
-        sistema->particulas[i].posicao = posicao;
-        sistema->particulas[i].velocidade = (Vector2){
-            (float)(GetRandomValue(-200, 200)),
-            (float)(GetRandomValue(-200, 200))};
-        sistema->particulas[i].cor = cor;
-        sistema->particulas[i].vida = 1.0f;
-        sistema->particulas[i].vidaMaxima = 1.0f;
-        sistema->particulas[i].tamanho = (float)GetRandomValue(3, 8);
-    }
-}
-
-void AtualizarParticulas(SistemaParticulas *sistema, float deltaTime)
-{
-    if (!sistema->ativo)
-        return;
-
-    sistema->timer += deltaTime;
-
-    for (int i = 0; i < sistema->numParticulas; i++)
-    {
-        Particula *p = &sistema->particulas[i];
-
-        p->posicao.x += p->velocidade.x * deltaTime;
-        p->posicao.y += p->velocidade.y * deltaTime;
-
-        p->velocidade.y += 300.0f * deltaTime;
-
-        p->vida -= deltaTime * 2.0f;
-
-        float alpha = p->vida / p->vidaMaxima;
-        p->cor.a = (unsigned char)(255 * alpha);
-    }
-
-    if (sistema->timer > 1.0f)
-    {
-        sistema->ativo = false;
-    }
-}
-
-void DesenharParticulas(SistemaParticulas *sistema)
-{
-    if (!sistema->ativo)
-        return;
-
-    for (int i = 0; i < sistema->numParticulas; i++)
-    {
-        Particula *p = &sistema->particulas[i];
-        if (p->vida > 0)
-        {
-            DrawCircleV(p->posicao, p->tamanho, p->cor);
-        }
-    }
-}
-
-void InicializarConfiguracaoPadrao(ConfiguracaoJogo *config)
-{
-
-    config->teclaPoderP1 = KEY_Q;
-    config->teclaSocoP1 = KEY_E;
-    config->teclaChute1 = KEY_R;
-    config->teclaEsquerdaP1 = KEY_A;
-    config->teclaDireitaP1 = KEY_D;
-    config->teclaConfirmarP1 = KEY_ENTER;
-
-    config->teclaPoderP2 = KEY_P;
-    config->teclaSocoP2 = KEY_O;
-    config->teclaChute2 = KEY_I;
-    config->teclaEsquerdaP2 = KEY_LEFT;
-    config->teclaDireitaP2 = KEY_RIGHT;
-    config->teclaConfirmarP2 = KEY_SPACE;
-
-    config->volumeMusica = 0.5f;
-    config->volumeEfeitos = 0.7f;
-}
-
-const char *ObterNomeTecla(int tecla)
-{
-    switch (tecla)
-    {
-    case KEY_Q:
-        return "Q";
-    case KEY_W:
-        return "W";
-    case KEY_E:
-        return "E";
-    case KEY_R:
-        return "R";
-    case KEY_T:
-        return "T";
-    case KEY_Y:
-        return "Y";
-    case KEY_U:
-        return "U";
-    case KEY_I:
-        return "I";
-    case KEY_O:
-        return "O";
-    case KEY_P:
-        return "P";
-    case KEY_A:
-        return "A";
-    case KEY_S:
-        return "S";
-    case KEY_D:
-        return "D";
-    case KEY_F:
-        return "F";
-    case KEY_G:
-        return "G";
-    case KEY_H:
-        return "H";
-    case KEY_J:
-        return "J";
-    case KEY_K:
-        return "K";
-    case KEY_L:
-        return "L";
-    case KEY_Z:
-        return "Z";
-    case KEY_X:
-        return "X";
-    case KEY_C:
-        return "C";
-    case KEY_V:
-        return "V";
-    case KEY_B:
-        return "B";
-    case KEY_N:
-        return "N";
-    case KEY_M:
-        return "M";
-    case KEY_SPACE:
-        return "SPACE";
-    case KEY_ENTER:
-        return "ENTER";
-    case KEY_LEFT:
-        return "←";
-    case KEY_RIGHT:
-        return "→";
-    case KEY_UP:
-        return "↑";
-    case KEY_DOWN:
-        return "↓";
-    default:
-        return "?";
-    }
-}
-
-void DesenharTelaOpcoes(ConfiguracaoJogo *config, int opcaoSelecionada, bool aguardandoTecla, Font fonte)
-{
-
-    DrawRectangle(0, 0, LARGURA_TELA, ALTURA_TELA, Fade(BLACK, 0.9f));
-
-    const char *titulo = "OPÇÕES";
-    Vector2 tamanhoTitulo = MeasureTextEx(fonte, titulo, 64, 2);
-    DrawTextEx(fonte, titulo, (Vector2){LARGURA_TELA / 2 - tamanhoTitulo.x / 2, 100}, 64, 2, WHITE);
-
-    const char *opcoes[OPCOES_CONFIG] = {
-        "CONTROLES PLAYER 1",
-        "CONTROLES PLAYER 2",
-        "VOLUME MÚSICA",
-        "VOLUME EFEITOS",
-        "RESETAR CONTROLES",
-        "SALVAR CONFIGURAÇÕES",
-        "CARREGAR CONFIGURAÇÕES",
-        "VOLTAR"};
-
-    int posY = 250;
-    int espacamento = 80;
-
-    for (int i = 0; i < OPCOES_CONFIG; i++)
-    {
-        Color corTexto = (i == opcaoSelecionada) ? YELLOW : WHITE;
-        Color corCaixa = (i == opcaoSelecionada) ? Fade(YELLOW, 0.3f) : Fade(WHITE, 0.1f);
-
-        int larguraCaixa = 800;
-        int alturaCaixa = 60;
-        int x = (LARGURA_TELA - larguraCaixa) / 2;
-        int y = posY + i * espacamento - 10;
-
-        DrawRectangleRounded((Rectangle){x, y, larguraCaixa, alturaCaixa}, 0.1f, 10, corCaixa);
-        DrawRectangleRoundedLines((Rectangle){x, y, larguraCaixa, alturaCaixa}, 0.1f, 10, corTexto);
-
-        DrawTextEx(fonte, opcoes[i], (Vector2){x + 20, posY + i * espacamento}, 36, 2, corTexto);
-
-        if (i == 0)
-        {
-            const char *controles = TextFormat("Poder:%s Soco:%s Chute:%s",
-                                               ObterNomeTecla(config->teclaPoderP1),
-                                               ObterNomeTecla(config->teclaSocoP1),
-                                               ObterNomeTecla(config->teclaChute1));
-            DrawTextEx(fonte, controles, (Vector2){x + 20, posY + i * espacamento + 35}, 24, 2, LIGHTGRAY);
-        }
-        else if (i == 1)
-        {
-            const char *controles = TextFormat("Poder:%s Soco:%s Chute:%s",
-                                               ObterNomeTecla(config->teclaPoderP2),
-                                               ObterNomeTecla(config->teclaSocoP2),
-                                               ObterNomeTecla(config->teclaChute2));
-            DrawTextEx(fonte, controles, (Vector2){x + 20, posY + i * espacamento + 35}, 24, 2, LIGHTGRAY);
-        }
-        else if (i == 2)
-        {
-            int barraLargura = 200;
-            int barraX = x + 400;
-            int barraY = posY + i * espacamento + 15;
-
-            DrawRectangle(barraX, barraY, barraLargura, 20, DARKGRAY);
-            DrawRectangle(barraX, barraY, (int)(barraLargura * config->volumeMusica), 20, GREEN);
-            DrawRectangleLines(barraX, barraY, barraLargura, 20, WHITE);
-
-            const char *volumeTexto = TextFormat("%.0f%%", config->volumeMusica * 100);
-            DrawTextEx(fonte, volumeTexto, (Vector2){barraX + barraLargura + 10, barraY}, 24, 2, WHITE);
-        }
-        else if (i == 3)
-        {
-            int barraLargura = 200;
-            int barraX = x + 400;
-            int barraY = posY + i * espacamento + 15;
-
-            DrawRectangle(barraX, barraY, barraLargura, 20, DARKGRAY);
-            DrawRectangle(barraX, barraY, (int)(barraLargura * config->volumeEfeitos), 20, BLUE);
-            DrawRectangleLines(barraX, barraY, barraLargura, 20, WHITE);
-
-            const char *volumeTexto = TextFormat("%.0f%%", config->volumeEfeitos * 100);
-            DrawTextEx(fonte, volumeTexto, (Vector2){barraX + barraLargura + 10, barraY}, 24, 2, WHITE);
-        }
-    }
-
-    if (aguardandoTecla)
-    {
-        const char *instrucao = "PRESSIONE UMA TECLA PARA CONFIGURAR...";
-        Vector2 tamanhoInstrucao = MeasureTextEx(fonte, instrucao, 28, 2);
-        DrawTextEx(fonte, instrucao, (Vector2){LARGURA_TELA / 2 - tamanhoInstrucao.x / 2, ALTURA_TELA - 100}, 28, 2, YELLOW);
-    }
-    else
-    {
-        const char *instrucoes = "↑/↓: Navegar | ENTER: Selecionar | ←/→: Ajustar Volume | BACKSPACE: Voltar";
-        Vector2 tamanhoInstrucoes = MeasureTextEx(fonte, instrucoes, 24, 2);
-        DrawTextEx(fonte, instrucoes, (Vector2){LARGURA_TELA / 2 - tamanhoInstrucoes.x / 2, ALTURA_TELA - 80}, 24, 2, LIGHTGRAY);
-    }
-}
-
-void SalvarConfiguracoes(ConfiguracaoJogo *config)
-{
-    // Implemente a lógica para salvar as configurações no arquivo
-    printf("Configurações salvas com sucesso!\n");
-}
-
-void CarregarConfiguracoes(ConfiguracaoJogo *config)
-{
-    // Implemente a lógica para carregar as configurações do arquivo
-    printf("Configurações carregadas com sucesso!\n");
-}
+#include "structs.h"
+#include "defines.h"
+#include "personagem.h"
+#include "ui.h"
+#include "efeitos.h"
+#include "config.h"
+
+// All helper functions have been moved to their respective .c files.
+// unifight.c now only contains the main game loop and initialization.
 
 int main(void)
 {
